+++ tests/integration/helpers/mongos.py
@@ -20,6 +20,7 @@ from tests.integration.helpers.common import (
     TIMEOUT,
     clear_continous_writes,
     deploy_charm,
+    external_cert_path,
     find_unit,
     get_address_of_unit,
     get_application_relation_data,
@@ -28,6 +29,7 @@ from tests.integration.helpers.common import (
     get_secret_data,
     get_unit_hostnames,
     get_unit_id,
+    internal_cert_path,
     mongosh,
     start_continous_writes,
     stop_continous_writes,
@@ -39,10 +41,8 @@ from tests.integration.helpers.tls import (
     cannot_connect_without_tls,
     check_certs_correctly_distributed,
     check_tls,
-    external_cert_path,
     get_file_content,
     integrate_apps_with_tls,
-    internal_cert_path,
     remove_tls_integrations,
     set_private_keys,
     time_file_created,
